+++ packages/help-center/src/styles.scss
@@ -8,6 +8,11 @@ $head-foot-height: 50px;
 	z-index: 9999;
 	cursor: default;
 
+	& > div {
+		display: flex;
+		flex-direction: column;
+	}
+
 	.help-center__container-header {
 		height: $head-foot-height;
 		padding: 16px 8px 16px 16px;
@@ -43,8 +48,8 @@ $head-foot-height: 50px;
 		}
 
 		.help-center__container-content {
+			flex-grow: 1;
 			height: calc( 80vh - $head-foot-height * 2 );
-			max-height: 700px;
 		}
 
 		&.is-minimized {
